+++ src/javascript/configuration.js
@@ -29,17 +29,20 @@ let current_config = { // Default values
 function loadConfig () { // Loads the configuration from the config.json for the first time
   if (!fs.existsSync(configDir)) {
     fs.writeFile(configDir, JSON.stringify(current_config)) // Save the config
-    detectPlugins()
     updateSettings()
     loadLanguage(current_config.language)
-    if (current_config.justInstalled === false) {
-      g_welcomePage()
-    } else {
-      Setup.open()
-    }
-    screens.add()
-    detectPlugins()
-    appendBinds()
+    
+    
+    detectPlugins(function(){
+      if (current_config.justInstalled === false) {
+        g_welcomePage()
+      } else {
+        Setup.open()
+      }
+      screens.add()
+      appendBinds()
+    })
+    
   } else {
     fs.readFile(configDir, 'utf8', function (err, data) {
       Object.keys(current_config).forEach(function (key, index) {